[WIP] Add support for basic service configuration
This change adds a file with environment variables for
each service that's provisioned into SNAP_DATA, which
allows it to be modifed by an admin (and eventually a
config hook).  Each service has an environment variable
that can be set to "y" or "n".  The start-edgex reads
this file, and only starts services for which the env
var is set to "y".